new designer-page

+++ Services/authenticateUser.js
@@ -1,9 +1,8 @@
-//Функция на авторизацию
-//для страниц: #2, 4
-
 import { auth, signInWithEmailAndPassword } from "../main.js";
+import { ref, get, child } from "firebase/database";
+import { database } from "../main.js";
 
-export default async function authenticateUser(email, password = null) {
+export default async function authenticateUser(email, password) {
   try {
     const userCredential = await signInWithEmailAndPassword(
       auth,
@@ -12,18 +11,27 @@ export default async function authenticateUser(email, password = null) {
     );
     const user = userCredential.user;
 
+    // Получение дополнительных данных пользователя из базы данных
+    const userRef = ref(database, `users/${user.uid}`);
+    const snapshot = await get(userRef);
+
+    if (!snapshot.exists()) {
+      throw new Error("User data not found in database.");
+    }
+
+    const userData = snapshot.val();
+
     // Сохраняем данные пользователя в Local Storage
     localStorage.setItem(
       "user",
       JSON.stringify({
         uid: user.uid,
-        name: user.name,
-        phoneNumber: user.userNumber,
+        name: userData.name,
+        phoneNumber: userData.phoneNumber,
         email: user.email,
-        password: user.password,
-        cart: user.cart,
-        currentOrders: user.currentOrders,
-        orderHistory: user.orderHistory,
+        cart: userData.cart,
+        currentOrders: userData.currentOrders,
+        orderHistory: userData.orderHistory,
         // другие данные, если нужно
       })
     );
